requirements.txt
hikari
hikari-lightbulb
hikari-miru

python-dotenv